pkg rename service -> server
