+++ cmd/translate/main_grpc_integration_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.expect.digital/translate/pkg/model"
 	translatev1 "go.expect.digital/translate/pkg/pb/translate/v1"
-	translateService "go.expect.digital/translate/pkg/service"
+	"go.expect.digital/translate/pkg/server"
 	"go.expect.digital/translate/pkg/testutil"
 	"golang.org/x/text/language"
 	"google.golang.org/genproto/protobuf/field_mask"
@@ -36,7 +36,7 @@ func randUploadData(t *testing.T, schema translatev1.Schema) ([]byte, language.T
 		messages.Messages = append(messages.Messages, message)
 	}
 
-	data, err := translateService.MessagesToData(schema, messages)
+	data, err := server.MessagesToData(schema, messages)
 	require.NoError(t, err, "convert rand messages to serialized data")
 
 	return data, lang
