+++ pkg/server/field_mask.go
@@ -1,4 +1,4 @@
-package service
+package server
 
 import (
 	"fmt"
