+++ pkg/server/file.go
@@ -1,4 +1,4 @@
-package service
+package server
 
 import (
 	"context"
