+++ pkg/server/file_test.go
@@ -1,4 +1,4 @@
-package service
+package server
 
 import (
 	"errors"
